+++ Plugins/WorldAssembly/Source/NexusWorldAssembly/Private/Assembly/Tasks/NSpawnCellProxiesTask.cpp
@@ -6,6 +6,7 @@
 
 void FNSpawnCellProxiesTask::DoTask(ENamedThreads::Type CurrentThread, const FGraphEventRef& CompletionGraphEvent)
 {
+	
 	N_ASSEMBLY_ANALYTICS_INDEX_DEFINE(SpawnCellProxiesCreate)
 	N_ASSEMBLY_ANALYTICS_INDEX(SpawnCellProxiesStart)
 	
@@ -28,6 +29,7 @@ void FNSpawnCellProxiesTask::DoTask(ENamedThreads::Type CurrentThread, const FGr
 	const double MaxAllowableTime = SpawnCellsContextPtr->CellTimeSlice;
 	const double StartTime = FPlatformTime::Seconds();
 	const int32 NodeCount = SpawnCellsContextPtr->CellNodes.Num();
+	TaskGraphContextPtr->SetStatusMessage(FString::Printf(TEXT("Spawning Cell Proxies (%i)"), NodeCount));
 	
 	// Copy local
 	const FGameplayTagContainer ContextTags = TaskGraphContextPtr->ContextTags;